+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this module will be documented in this file.
 ### Added
 - Results count has been added to keep track of the total results of a search term
 
+### Fixed
+- Make sure search results show max 25 results when loading collection for first time 
+
 ## 1.0.2 [22-05-2025]
 ### Added
 - Enable/Disable ajax search logging. It's disabled by default